refactor(phase-5): dedup + extraction — single homes, disposal, debug-gate hook
Behaviour-preserving quality pass (docs/refactor-plan.md Phase 5):
- src/lib/motion.ts: EASE token (was declared in 7 overlay/nav files)
- synty/{demoInstances,materials,textures}.ts: demoData (cast ×4), applyEmissive
  (×2), useEmissiveAtlas (×2) each get one home
- three/shaders/noise.ts: NOISE_GLSL spliced into the loader-veil dissolve and
  atmosphere-fog shaders (identical hash/vnoise pair) — verified pixel-identical
- three/shared/glowTexture.ts: getGlowTexture moved out of the Synty translation
  layer (it's a generic halo used by bulbs/moon/tickets/striker); 4 importers repointed
- three/hooks/useDebugGates.ts: Scene's ?off/?offp/?bloom URL gates extracted
- three/hooks/useDisposable.ts (useDisposeOnUnmount): disposes GPU resources on
  unmount — fixes LoaderVeil's real per-visit ShaderMaterial+geometry leak; also
  Atmosphere + GoldenTickets. NOT the plan's useDisposable(factory,deps) shape:
  the repo's react-hooks rule hard-errors on non-literal useMemo deps, so the
  hook takes an already-created resource instead (deviation recorded).

Gate: typecheck/lint(0 warnings)/42 tests/build green; overview + loader veil
screenshots pixel-identical; dedup greps confirm single homes.
