+++ src/components/three/shaders/noise.ts
@@ -1,0 +1,21 @@
+/**
+ * Cheap value noise (hash + bilinear-smoothed lattice) for GLSL fragment
+ * shaders — splice into a shader source before the function that uses `vnoise`.
+ * Shared by the loader veil dissolve and the atmosphere fog so the two can't
+ * drift apart.
+ */
+export const NOISE_GLSL = /* glsl */ `
+float hash(vec2 p) {
+  return fract(sin(dot(p, vec2(127.1, 311.7))) * 43758.5453123);
+}
+float vnoise(vec2 p) {
+  vec2 i = floor(p);
+  vec2 f = fract(p);
+  vec2 u = f * f * (3.0 - 2.0 * f);
+  return mix(
+    mix(hash(i), hash(i + vec2(1.0, 0.0)), u.x),
+    mix(hash(i + vec2(0.0, 1.0)), hash(i + vec2(1.0, 1.0)), u.x),
+    u.y
+  );
+}
+`;
