+++ src/components/three/synty/SyntyScene.tsx
@@ -1,11 +1,12 @@
 import { useGLTF, useTexture } from '@react-three/drei';
-import { LinearFilter, SRGBColorSpace, type Texture } from 'three';
+import { SRGBColorSpace, type Texture } from 'three';
 
-import demoInstances from './demo-instances.json';
 import manifest from './manifest.json';
 import { InstancedPrefab } from './InstancedPrefab';
 import { SPINNER_NAMES } from './AnimatedRides';
 import { EXTRA_INSTANCES } from './sceneAdditions';
+import { useEmissiveAtlas } from './textures';
+import { demoData } from './demoInstances';
 
 const SYNTY = '/models/synty/';
 const available = new Set(manifest as string[]);
@@ -35,7 +36,7 @@ const inPile = (name: string, t: number[]) =>
   !PILE_KEEP.test(name) && (t[0] - PILE_X) ** 2 + (-t[2] - PILE_Z) ** 2 < PILE_R2;
 
 /** [prefabName, transforms[]] for every demo prefab we have a GLB for, plus our additions. */
-export const entries = Object.entries(demoInstances as Record<string, number[][]>)
+export const entries = Object.entries(demoData)
   .filter(([name]) => available.has(name) && !EXCLUDE.test(name) && !SPINNER_NAMES.has(name))
   .map(([name, tfs]) => {
     const kept = tfs.filter((t) => !inPile(name, t));
@@ -61,21 +62,7 @@ export function SyntyScene({
   /** Debug-only (headless bisection): drop prefabs whose name matches. */
   exclude?: RegExp;
 }) {
-  // configure in the load callback (glTF UV convention + sRGB glow); mostly-black
-  // map so non-neon pixels stay dark.
-  const emissive = useTexture(`${SYNTY}emissive-atlas.png`, (t) => {
-    const tex = t as Texture;
-    tex.flipY = false;
-    tex.colorSpace = SRGBColorSpace;
-    // The Synty emissive atlas is 99% black with the bulb/light swatches packed
-    // into a tiny lit corner. With mipmaps on, those few-texel swatches average to
-    // black as the rides recede, so the bulbs go dark — disable mipmaps so every
-    // bulb samples the sharp lit texel at any distance (light twinkle is a bonus).
-    tex.generateMipmaps = false;
-    tex.minFilter = LinearFilter;
-    tex.magFilter = LinearFilter;
-    tex.needsUpdate = true;
-  }) as Texture;
+  const emissive = useEmissiveAtlas();
 
   // Shared base atlas — re-textures props the conversion left blank (most use it).
   const baseAtlas = useTexture(`${SYNTY}base-atlas.png`, (t) => {
